feat(gtk): Display upgrade dialog with changelog before reboot
